Refactor that enables lein uberjar to complete
In particular, the prone.middleware namespace needed to be moved to dev only (now in
lambda-web.config for env/dev). For this reason, following the Luminous framework, a custom
limbdax-web.handler has been created which contains ring handler/middlewares.

A :pre-middleware and :post-middleware keys are now passed to the server component in order to
enrich the main app handler with more stuff.