Remove ^:replace "dev-resources" because Figwheel needs "resources"
There was a 404 that I could not understand but then I saw that I was replacing the :resource-paths
option for the dev profile. Of course Figwheel was not happy.

In the process, the deps has been updated, the cljs build id changed to "website" and the classic
"fig-dev" lein command added.